requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
gunicorn==21.2.0
python-dotenv==1.0.0
psutil==5.9.6
click